# support ancient versions of docker again (template parsing issue)