@afbjorklund 2021-02-05
I don't think we need to support Docker 17.xx or earlier, only OS that I know of that has that is CentOS 7 ([1.13.1](https://docs.docker.com/engine/release-notes/prior-releases/#1131-2017-02-08)).

And we **don't** want to support that, if you want to run CentOS you **need** to install a newer version from vendor:

https://docs.docker.com/engine/install/centos/

For Ubuntu, the oldest that we need to support is 18.09.7: https://packages.ubuntu.com/search?keywords=docker.io